refactor(#110): 파일 모듈화
