+++ apps/web/src/components/notfound/NotFoundContent/index.ts
@@ -1,0 +1,1 @@
+export { default as NotFoundContent } from './NotFoundContent';
