+++ apps/web/src/components/notfound/SlotMachineDigits/index.ts
@@ -1,0 +1,1 @@
+export { default as SlotMachineDigits } from './SlotMachineDigits';
