+++ apps/web/src/components/notfound/index.ts
@@ -1,0 +1,2 @@
+export { default as NotFoundContent } from './NotFoundContent/NotFoundContent';
+export { default as SlotMachineDigits } from './SlotMachineDigits/SlotMachineDigits';
